+++ src/support/serenity/serenity-transport-error.js
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2026 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/**
+ * Error thrown when the Semrush upstream returns a non-2xx response or refuses
+ * the auth header. `status` carries the upstream status; `body` is the parsed
+ * JSON (or raw text when not valid JSON). Controllers must not leak `.body` to
+ * clients — it is kept here only for server-side logging and classification.
+ */
+export class SerenityTransportError extends Error {
+  /**
+   * @param {number} status - the upstream HTTP status.
+   * @param {string} message
+   * @param {any} [body] - parsed JSON, raw text, or null for an empty body.
+   */
+  constructor(status, message, body) {
+    super(message);
+    this.name = 'SerenityTransportError';
+    this.status = status;
+    this.body = body;
+  }
+}
